+++ Changelog.md
@@ -1,6 +1,8 @@
 # Changelog
 
 ## 2024-24-01
+- Display a space invader in the top left corner
+    - use minifb
 - Add function to dump all memory
 - Load the rom into chip memory
 - Read a rom as parameter and print opcode